+++ routes/general.routes.js
@@ -15,6 +15,7 @@ const discapacityRoutes = require("./discapacity.routes");
 const disabilitiesRoutes = require('./disabilities.routes');
 const notificationsRoutes = require('./notifications.routes');
 const attendanceRoutes = require('./attendance.routes');
+const uploadsRoutes = require('./uploads.routes');
 const {authenticateToken} = require("../utils/auth.middleware");
 
 router.use(authenticateToken);
@@ -34,5 +35,6 @@ router.use("/disabilities", disabilitiesRoutes);
 router.use("/notifications", notificationsRoutes);
 router.use("/attendance", attendanceRoutes);
 
+router.use("/uploads", uploadsRoutes);
 
 module.exports = router;